Renamed model and discovery packages to internal in the Systeminfo binding. See openhab/static-code-analysis#100
