+++ addons/binding/org.openhab.binding.systeminfo.test/src/test/groovy/org/openhab/binding/systeminfo/test/SysteminfoOSGiTest.groovy
@@ -49,7 +49,7 @@ import org.junit.Ignore
 import org.junit.Test
 import org.junit.experimental.categories.Category
 import org.openhab.binding.systeminfo.SysteminfoBindingConstants
-import org.openhab.binding.systeminfo.discovery.SysteminfoDiscoveryService
+import org.openhab.binding.systeminfo.internal.discovery.SysteminfoDiscoveryService
 import org.openhab.binding.systeminfo.handler.SysteminfoHandler
 import org.openhab.binding.systeminfo.internal.SysteminfoHandlerFactory
 /**
