+++ addons/binding/org.openhab.binding.systeminfo/src/main/java/org/openhab/binding/systeminfo/internal/model/OshiSysteminfo.java
@@ -6,7 +6,7 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.systeminfo.model;
+package org.openhab.binding.systeminfo.internal.model;
 
 import java.math.BigDecimal;
